// app/src/app/store/IATI/IATIWriter/types.ts
interface FinanceProps {
  value: number
  status: number
  start: string
  end: string
}

export interface IATIBudgetProps {
  budgetType: number
  budgetLine: string
  otherRef: string
  finance: FinanceProps
}

export const enum IATIWriterActionTypes{
  TX_INIT =  '@@IATIWriterActionTypes/TX_INIT',
  ORGS_SUCCESS = '@@IATIWriterActionTypes/WRITEORGS_SUCCESS',
  ORGS_FAILURE = '@@IATIWriterActionTypes/WRITEORGS_FAILURE',
  ORGANISATIONS_SUCCESS = '@@IATIWriterActionTypes/WRITEORGANISATIONS_SUCCESS',
  ORGANISATIONS_FAILURE = '@@IATIWriterActionTypes/WRITEORGANISATIONS_FAILURE',
  ORGANISATION_SUCCESS = '@@IATIWriterActionTypes/WRITEORGANISATION_SUCCESS',
  ORGANISATION_FAILURE = '@@IATIWriterActionTypes/WRITEORGANISATION_FAILURE',
  BUDGET_SUCCESS = '@@IATIWriterActionTypes/WRITEBUDGET_SUCCESS',
  BUDGET_FAILURE = '@@IATIWriterActionTypes/WRITEBUDGET_FAILURE',
  RECIPIENTORGBUDGET_SUCCESS = '@@IATIWriterActionTypes/WRITERECIPIENTORGBUDGET_SUCCESS',
  RECIPIENTORGBUDGET_FAILURE = '@@IATIWriterActionTypes/WRITERECIPIENTORGBUDGET_FAILURE',
  RECIPIENTREGIONBUDGET_SUCCESS = '@@IATIWriterActionTypes/WRITERECIPIENTREGIONBUDGET_SUCCESS',
  RECIPIENTREGIONBUDGET_FAILURE = '@@IATIWriterActionTypes/WRITERECIPIENTREGIONBUDGET_FAILURE',
  RECIPIENTCOUNTRYBUDGET_SUCCESS = '@@IATIWriterActionTypes/WRITERECIPIENTCOUNTRYBUDGET_SUCCESS',
  RECIPIENTCOUNTRYBUDGET_FAILURE = '@@IATIWriterActionTypes/WRITERECIPIENTCOUNTRYBUDGET_FAILURE',
  TOTALEXPENDITURE_SUCCESS = '@@IATIWriterActionTypes/WRITETOTALEXPENDITURE_SUCCESS',
  TOTALEXPENDITURE_FAILURE = '@@IATIWriterActionTypes/WRITETOTALEXPENDITURE_FAILURE',
  DOCUMENT_SUCCESS = '@@IATIWriterActionTypes/WRITEDOCUMENT_SUCCESS',
  DOCUMENT_FAILURE = '@@IATIWriterActionTypes/WRITEDOCUMENT_FAILURE'
}